Port the remaining old-style ElasticIndexerTest
