+++ common/pipeline_storage/src/test/scala/uk/ac/wellcome/pipeline_storage/elastic/ElasticIndexerTest.scala
@@ -3,12 +3,13 @@ package uk.ac.wellcome.pipeline_storage.elastic
 import com.sksamuel.elastic4s.ElasticDsl._
 import com.sksamuel.elastic4s.{Index, Response}
 import com.sksamuel.elastic4s.requests.get.GetResponse
+import com.sksamuel.elastic4s.requests.mappings.dynamictemplate.DynamicMapping
 import org.scalatest.Assertion
-import uk.ac.wellcome.elasticsearch.NoStrictMapping
+import uk.ac.wellcome.elasticsearch.{IndexConfig, IndexConfigFields, NoStrictMapping, WorksAnalysis}
 import uk.ac.wellcome.elasticsearch.test.fixtures.ElasticsearchFixtures
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.json.JsonUtil.toJson
-import uk.ac.wellcome.pipeline_storage.fixtures.SampleDocument
+import uk.ac.wellcome.pipeline_storage.fixtures.{SampleDocument, SampleDocumentData}
 import uk.ac.wellcome.pipeline_storage.{ElasticIndexer, Indexer, IndexerTestCases}
 
 import scala.concurrent.ExecutionContext.Implicits.global
@@ -61,9 +62,54 @@ class ElasticIndexerTest
 
       val getResponse = response.result
 
-      getResponse.exists shouldBe true
+      // If there's a document with this ID, we want to make sure it's something
+      // different.  If there's no document with this ID, then all is well.
+      if (getResponse.exists) {
+        assertJsonStringsAreDifferent(getResponse.sourceAsString, documentJson)
+      } else {
+        assert(true)
+      }
+    }
+  }
 
-      assertJsonStringsAreDifferent(getResponse.sourceAsString, documentJson)
+  it("returns a list of documents that weren't indexed correctly") {
+    val validDocuments = (1 to 5).map { _ => createDocument }
+    val invalidDocuments = (1 to 3).map { _ =>
+      createDocument
+        .copy(data = SampleDocumentData(Some(randomAlphanumeric())))
+    }
+
+    object StrictWithNoDataIndexConfig
+      extends IndexConfig
+        with IndexConfigFields {
+
+      import com.sksamuel.elastic4s.ElasticDsl._
+
+      val analysis = WorksAnalysis()
+
+      val title = textField("title")
+      val data = objectField("data")
+
+      val mapping = properties(Seq(title, canonicalId, version, data))
+        .dynamic(DynamicMapping.Strict)
+    }
+
+    withLocalElasticsearchIndex(config = StrictWithNoDataIndexConfig) { implicit index =>
+      withIndexer { indexer =>
+        val future = indexer.index(validDocuments ++ invalidDocuments)
+
+        whenReady(future) { result =>
+          result.left.get should contain only (invalidDocuments: _*)
+
+          validDocuments.foreach { doc =>
+            assertIsIndexed(doc)
+          }
+
+          invalidDocuments.foreach { doc =>
+            assertIsNotIndexed(doc)
+          }
+        }
+      }
     }
   }
 }
